+++ test/runtests.jl
@@ -6,6 +6,7 @@
 using Test
 
 include("aqua.jl")
+include("explicit_imports.jl")
 include("test_core.jl")
 #include("test_glm_integration.jl")  # Requires RCall - use test_glm_integration_julia.jl instead
 include("test_glm_integration_julia.jl")  # Pure Julia GLM tests
@@ -18,6 +19,7 @@ include("test_varhac_fortran_validation.jl")
 include("test_smoothed_moments.jl")
 include("test_ewc_montecarlo.jl")
 include("test_types_coverage.jl")
+include("test_equality.jl")
 include("test_edge_cases.jl")
 include("test_api_coverage.jl")